Move test function get_only_file
Also remove o3 requirement, as this function will be useful for v2
writer/reader tests.
